pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slack-rag-bot"
version = "0.1.0"
description = "A Slack RAG bot with FastAPI and Docket for background processing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "slack-bolt>=1.14.0",
    "openai>=1.0.0",
    "redisvl>=0.6.0,<0.8.0",
    "sentence-transformers>=2.2.2",
    "redis>=4.5.0",
    "pydocket>=0.7.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "prefect>=3.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "nbformat>=5.9.0",
    "nbconvert>=7.0.0",
    "PyPDF2>=3.0.0",
    "markdownify>=0.11.6",
    "gitpython>=3.1.0",
    "reportlab>=4.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langchain>=0.3.25",
    "pytest-asyncio>=0.26.0",
    "aiofiles>=24.1.0",
    "httpx>=0.25.0",
    "tavily-python>=0.7.5",
    "click>=8.0.0",
    "glean-api-client>=0.6.0",
    "agent-memory-client>=0.9.2",
    "boto3>=1.34.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.41b1",
    "opentelemetry-instrumentation-openai>=0.40.14",
    "opentelemetry-exporter-otlp-proto-http>=1.34.0",
    "authlib>=1.6.3",
    "jinja2>=3.1.6",
    "langchain-community>=0.3.25",
    "pypdf>=6.0.0",
]

[project.scripts]
haink = "haink:cli"

[tool.hatch.build.targets.wheel]
# Includes all files in the root directory. Use only-include to narrow.
include = [
    "*.py",
    "*.md",
    "*.txt",
    "*.yaml",
    "*.yml",
    "*.json",
    "*.toml",
    "*.ini",
    "*.sh",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.3.0",
    "testcontainers>=3.7.0",
    "pre-commit>=3.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = ["-v", "--tb=short"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
